requirements.txt
pandas
numpy
azure-functions
semantic_kernel
langchain
openai
python-dotenv
azure-search-documents
ipykernel

